pyproject.toml
[tool.poetry]
name = "pocket_monster_world"
version = "0.1.0"
description = ""
authors = ["Bartosz 'b4rti' Gorol <bartgorol@t-online.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.0"
pygame-ce = "^2.1.3"
numpy = "^1.24.2"

[tool.poetry.scripts]
fetch-rest = "scripts.fetch_rest:run"
fetch-gql = "scripts.fetch_gql:run"
build = "scripts.build:run"
pmw = "pocket_monster_world.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
